backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
bcrypt
passlib[bcrypt]
pyjwt
pydantic
aiofiles
email-validator
pymongo
python-jose[cryptography]
**requests**
